feat(ui): add support for configurable side panels
Previously, only two side panels were supported, the layer side panel
and the selection details side panel.  Only a static configuration was
supported, a single column on the right, with the layer side panel on
top and the selection details below, and the vertical height divided
equally.

This commit adds support for an arbitrary number of side panels, to
the left/right/top/bottom of the main data panels.  Panels can be
moved via drag and drop, and the size can also be adjusted.  The side
panels for multiple layers can now be shown at once, and individual
tabs of side panels can also be dragged to separate panels.