feat: add layer list panel and support for "archived" layers
Archived layers are shown only in the new layer list panel; they are
not visible and are not shown in the layer bar.  They may be useful in
the case where there are a large number of layers that would otherwise
clutter up the layer bar.

This commit also improves the visual feedback during layer and drop
operations.