fix(ui/side_panel): only allow dragging from title bar
This allows slider controls in the body of side panels to work correctly.

Also disables auto-select on focus behavior for the name input field
of the layer side panels.  The previous auto-select on focus behavior
meant that attempting to start a drag from the text area (which takes
up most of the title bar and is the most natural place to drag from)
worked, but showed the text itself being dragged rather than the
entire panel.