fix(annotation/bounding_box): ensure bounding box dimensions != 0
Cross-section views do not display bounding box annotations where the
"depth" relative to the cross-section is 0.

That is partly an implementation artifact but also logically
consistent.  Previously, interactively creating an in-plane bounding
box annotation would produce an annotation of depth 0, which was not
visible.

With this change, creating a bounding box interactively ensures all
dimensions are at least 1, which ensures the bounding box being
created in-plane is visible.