feat(layer_bar): control+click/right-click unconditionally shows side panel
In e640fa4ba79a3de6c80ed135019a1c877f841db0, the behavior of the layer
bar was changed such that control+click on a layer would toggle the
visibility of the side panel, rather than unconditionally showing it.
However, it turns out this does not seem to provide a great UX; in
this commit, the behavior is changed back.