feat: support optional modifiers in event bindings
Events can now be specified with optional modifiers,
e.g. "shift?+alt?+keyx", which is equivalent to binding the event to
"keyx", "shift+keyx", "alt+keyx", and "alt+shift+keyx".

This is particularly useful when the same event bindings may be used
with or without a shift+letter tool binding being held down.