refactor(python): Reimplement communication to use server-sent events
Previously, sockjs was used for bidirectional state-updates and also
for credentials.  When websocket is available sockjs works fairly
well, but the fallbacks are not always reliable; in particular, it
does not work on colab.

Additionally, the bidirectional channel provided by sockjs was not
particularly well suited for obtaining credentials.

With this change, regular HTTP requests are used to obtain credentials
and send client -> Python state updates.

Python -> client state updates are sent using server-sent events.

This eliminates the client dependency on sockjs and the Python
dependency on sockjs-tornado, and allows Neuroglancer to work on
colab.