fix(s3): use s3 domains without region
Previously, Neuroglancer used a complicated mechanism to attempt to
determine the region of a bucket, but this mechanism did not work.
However, it turns out it was unnecessary because all buckets can
instead be accessed using `<bucket>.s3.amazonaws.com` regardless of
region.

Fixes https://github.com/google/neuroglancer/issues/418