feat(coordinate_space): support "integer coordinates specify voxel centers"
Previously, Neuroglancer maps integer coordinates to the corners of a
voxel by default, and the position snapping/rounding behavior assumed
this to be the case.

This rounding ensures that by default, the position does not land
exactly on the boundary between two voxels, as in that case it is not
defined which cross section should be displayed.

Users may wish to work with coordinate spaces using the "integer
coordinates specify voxel centers" convention, such as is the default
with OME-zarr.  To support that, Neuroglancer now automatically
decides for each dimension which convention is in use, using the
following heuristic:

- If for a given dimension at least one source has bounds exactly at
  half-integer coordinates, and no source has bounds at integer
  coordinates, assume the "integer coordinates specify voxel centers"
  convention.

- Otherwise, assume the "half-integer coordinates specify voxel
  centers" convention.