fix: fix intermittent stalling/failing of chunk priority updates
This fixes the longstanding bug whereby, seemingly at random, new chunks
would no longer load until layer visibility is toggled.

The issue was due to order dependence in signal dispatching in response
to changes to the DisplayDimensionRenderInfo. Previously,
DisplayDimensionRenderInfo was synchronized between the main UI thread
and the chunk worker thread in a fragile way: the transformed sources
computed for a layer in the UI thread were implicitly associated with a
particular `DisplayDimensionRenderInfo` value, but this value was not
transmitted with the sources, and instead was relied on being
synchronized independently.

This commit instead ensures that the associated
`DisplayDimensionRenderInfo` value is always sent along with the
transformed sources, which avoids this order dependence.