fix(coordinate_space): handle nearly integer-aligned bounds
Previously, coordinate bounds approximately, but not exactly, equal to
an integer or half-integer were handled poorly.

In particular, the lower/upper bounds could end up off by 1 from the
expected value, and voxelCenterAtIntegerCoordinates could end up false
when it was expected to be true, leading to coordinates snapping to the
boundary between two voxels, rather than the center, as expected.

With this change, coordinate bounds within 1e-3 of an integer or
half-integer are handled properly.

Fixes https://github.com/google/neuroglancer/issues/719.