fix(ui/side_panel): ensure panel cross sizes are correctly restored from JSON state (#774)
A given panel "flex group" (column/row) may contain both visible and non-visible
panels.

Previously, the presence of a non-visible panel could result in the size not
being correctly restored from the JSON state.

With this change:

- The size of the flex group is taken from the first visible panel, rather than
the first visible or non-visible panel.

- When the first visible panel is added to a flex group that was previously
hidden (due to not containing any visible panels), its size is set to the size
of the newly visible panel.  This means that two panels in the same flex group
can maintain different sizes (e.g. the "settings" and "layer list" panels), as
long as both are not visible at the same time.  If both become visible at the
same time, the second one to become visible will have its size reset to the size
of the other panel.

Fixes https://github.com/google/neuroglancer/issues/773.