chore: update npm deps
tsx is no longer used for the build scripts, as a subset of typescript is
supported natively by node>=22.18.

Lint and typecheck are no longer included in build_tools/cli.  These were
previously enabled by default, but in most cases it was inconvenient to have
these steps slowing down the build.  Additionally, the existing integrations did
not support tsgo or oxlint.